fix bugs x5

+++ landscape/src/main/resources/db/alter-table-account.sql
@@ -3,7 +3,7 @@
 --changeset bnkspbrus:2
 create table account_type_v2
 (
-    id        int,
+    id        int8,
     type_name varchar(255) unique not null,
     primary key (id)
 );
@@ -13,4 +13,4 @@ values (1, 'handyman'),
        (2, 'rancher');
 
 alter table account
-    add type_v2 int references account_type_v2 (id) not null;
+    add type_v2 int8 references account_type_v2 (id) not null;
